requirements.txt
requests
PyYAML
pandas
tqdm
pytrends
beautifulsoup4
google-genai
